Split up TS files
